requirements.txt
asgiref==3.2.10
Django==3.1.1
django-nested-inline==0.4.2
pytz==2020.1
sqlparse==0.3.1
